pyproject.toml
[tool.poetry]
name = "personal project"
version = "0.1.0"
description = ""
authors = ["Duc Nguyen <nhdnguyen21@clc.fitus.edu.vn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
openai = "^1.56.2"
numpy = "^2.1.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
